requirements.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
pyinstaller==6.3.0 